NSS-backend/requirements.txt
Django>=5.1.5
djangorestframework
django-cors-headers
psycopg2-binary
gunicorn
whitenoise
python-dotenv
